Update queen_attack_challenge.py
Purpose of this is to see how many spaces the queen on a chess board of size n with number k obstacles blocking it's path.  Problem taken from Hackerrank
https://www.hackerrank.com/challenges/queens-attack-2/problem?isFullScreen=true
Currently, some of the test cases have timed out of time limit when testing, so a more elegant or efficient solution should be required to get those couple final test cases.  I suspect breaking out some linear algebra will allow the solution to become much more efficient computationally.  The current method is crude and brute force; it checks each of the eight paths the queen could go, walking along the path until a obstacle or edge of the board is encountered, and it then jumps to the next path.  For a normal sized chess board this would be fine, but for one that is up to 100 000^2 total squares with 100 000 obstacles could take some time.  These test cases are locked behind a paywall, so I can't be fully sure why they don't work for now.